using vite to build /ui app
