+++ .nvmrc
@@ -1,0 +1,1 @@
+14.16.0
