+++ ui/package.json
@@ -27,6 +27,9 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
+    "vite":"vite",
+    "vite-build":"vite build",
+    "vite-preview":"vite preview",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
@@ -47,5 +50,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "vite": "^2.2.4"
   }
 }
