+++ ui/src/pages/register/index.jsx
